requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
openai>=1.1.0